% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avg_pop}
\alias{avg_pop}
\title{Average population by district}
\format{A data frame with 120 rows and 3 variables:
\itemize{
\item \code{district}: name of the district without accent
\item \code{year}: year
\item \code{avrg_pop}: average population by number of people
}}
\source{
The District Statistics Office of Ho CHi Minh
\url{http://www.pso.hochiminhcity.gov.vn/web/guest/home}
}
\usage{
avg_pop
}
\description{
A dataset containing the average population by number of person for Ho Chi
Minh Province from 2011 to 2015 in a tidy version of the raw data coming from
the District Statistics Office of Ho CHi Minh, from the section "Population
and labour", table "Average population by district".
}
\keyword{datasets}
